Avoid printing timer and statistic data manually
Previously in MCABridge we need to print them manually because QEMU --
or shared libraries in general -- seems to have a slightly different
finialization process so LLVM's print-on-exit mechanism didn't work so
well in that scenario.
But now we don't have that problem when it comes to executable. So
LLVM's built-in `-stats` and default timer printing behavior should be
sufficient.