Fix the issue that creates indeterministic memory access MD
Instead of using MCInst* as the key to IndexMap in MDExchanger, which
causes problem when the fetched block is consisting of multiple TBs,
we're using index of MCInst in the fetched block.
In addition to this, we're not checking if the result from
llvm::lower_bound (std::lower_bound) is end() anymore, since it's
totally possible that the new TBSlice will take all the memory access
entries.