Add a new BinaryRegions operation mode
The "marker" operation mode serves the same purpose of the original
sub-region feature, but instead of chopping instruction stream, it
merely marked the begin / end instruction so that MCA view can perform
special handling on the marked region. For example, our version of
SummaryView will print the summary upon encountering either the begin or
end marker, such that we can retrieve an accurate cycle count -- in a
context-sensitive fashion -- of certain program region.