WIP MCAD Vivisect broker
This currently seems to work on an x86 binary. The MCAD
Vivisect Broker acts as a gRPC client to the Vivisect gRPC
server. Both must be started separately, with the server
getting started first (one must wait for it to completely
analyze the given binary). MCAD calls the broker's fetch
method which then invokes the functionality of the
Vivisect server and in return receives an array of
instructions that were emulated. Those instructions
have raw portions that get converted to an array of LLVM
`MInst`. Those are populated in `fetch` and the total
number of instruction ran is returned. This allows MCAD
to evaluate the batch. Currently, all instructions are
run at once.